journal-frontend/backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
pydantic
python-dotenv
psycopg2-binary
